Restructuring Scripts
MapCollision- checks whether the player is hovering over a collider and has the public boolean isValidLocation, we can use this is our other scripts.

Player Movement- removed the boolean isValidLocation and moved it to MapCollision script, now purely controls the player movement. now also takes in the player in the fields so we can move this script to the game manager.

CursorController- the new and improved version of the now removed CursorManager script. Now it only changes the cursor and will do so for all our future scenarios.